:recycle: refactor: delegate package 수정
Issue: #24
