+++ app/src/main/java/com/bibbidi/habittracker/ui/common/dialog/EmojiPickerBottomSheet.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.view.View
 import com.bibbidi.habittracker.R
 import com.bibbidi.habittracker.databinding.BottomSheetInputEmojiBinding
-import com.bibbidi.habittracker.ui.common.viewBinding
+import com.bibbidi.habittracker.ui.common.delegate.viewBinding
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.vanniktech.emoji.listeners.OnEmojiClickListener
