+++ app/src/main/java/com/bibbidi/habittracker/ui/home/HabitsAdapter.kt
@@ -11,7 +11,7 @@ import com.bibbidi.habittracker.databinding.ItemHabitCheckBinding
 import com.bibbidi.habittracker.databinding.ItemHabitTimeBinding
 import com.bibbidi.habittracker.databinding.ItemHabitTrackBinding
 import com.bibbidi.habittracker.ui.common.BaseViewHolder
-import com.bibbidi.habittracker.ui.common.viewBinding
+import com.bibbidi.habittracker.ui.common.delegate.viewBinding
 import com.bibbidi.habittracker.ui.model.habit.log.CheckHabitLogUiModel
 import com.bibbidi.habittracker.ui.model.habit.log.HabitLogUiModel
 import com.bibbidi.habittracker.ui.model.habit.log.TimeHabitLogUiModel
